⚙️chore(CHEC-52): 패키지 이름을 entity에서 domain으로 변경
