+++ src/main/java/com/checkmate/bub/speech/domain/BookmarkEntity.java
@@ -1,4 +1,4 @@
-package com.checkmate.bub.speech.entity;
+package com.checkmate.bub.speech.domain;
 
 import com.checkmate.bub.global.config.audit.BaseTimeEntity;
 import jakarta.persistence.*;
@@ -8,16 +8,16 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * AR-ADMIN-005: 개인화된 문장 저장 엔티티
- * 사용자의 커스텀 문장을 저장하고 관리
+ * UR-USER-021, UR-USER-022, AR-ADMIN-005: 확언 북마크 엔티티
+ * 사용자가 문장을 정확히 읽은 후 북마크에 추가/제거할 수 있는 기능
  */
 @Entity
-@Table(name = "custom_sentences")
+@Table(name = "bookmarks")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class CustomSentenceEntity extends BaseTimeEntity {
+public class BookmarkEntity extends BaseTimeEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -27,21 +27,11 @@ public class CustomSentenceEntity extends BaseTimeEntity {
     @Column(nullable = false)
     private String userNickname;
     
-    // 커스텀 문장 내용
+    // 북마크된 문장 내용
     @Column(nullable = false, length = 1000)
     private String sentence;
     
     // 문장의 톤
     @Column(nullable = false)
     private String tone;
-    
-    // 문장 수정을 위한 메서드
-    public void updateSentence(String newSentence) {
-        this.sentence = newSentence;
-    }
-    
-    // 톤 변경을 위한 메서드
-    public void updateTone(String newTone) {
-        this.tone = newTone;
-    }
 }
